tigris statics: split out bucket finding into FindBucket
